==> Code/sandbox/Main(2024)/3_Make_figures_tables.Rmd <==
---
title: "3_Make_figures_tables"
author: "Jaeseok Hwang"
date: "`r Sys.Date()`"
output:
  html_document: default
  pdf_document: default
  word_document: default
---


## Knitr option

```{r, cache = F, echo = F, results = "hide"}
#####

library(knitr)

knitr::opts_chunk$set(
  cache = FALSE,
  echo = FALSE,
  warning = FALSE,
  cache.lazy = FALSE,
  fig.retina = 6,
  fig.height = 9,
  fig.width = 9,
  message = FALSE,
  error = TRUE
)

options(knitr.duplicate.label = "allow")

```


#### Packages 

```{r pacakages, cache = FALSE, results = "hide"}

library(here)
library(rmarkdown) # for rmarkdown options
library(bookdown) # for bookdown options
library(knitr) # for knitr options

library(sf) # for spatial data

library(data.table) # for data manipulation
library(tidyverse) # for data manipulation
library(dplyr) # for data manipulation
library(lubridate) # for date manipulation

library(tmap) # for mapping
library(ggplot2) # for plotting
library(kableExtra) # for table formatting
library(webshot) # for converting HTML to PNG
library(patchwork)
library(flextable)

```

### Read sources and load processed and analysis data

```{r source, echo = F, results = "hide"}

# Read functions for data processing 
source(here("Code","Main","0_Set_up_preparation.R"))
source(here("Code","Functions","functions_for_table_figure.R"))

# 101 field data
info_tb_list <-readRDS(here("Data", "Processed", "Analysis_results", "info_tb_list.rds"))

# Convert the list to a single data.table and remove the input_type and unit columns

info_tb_101 <- bind_rows(info_tb_list) %>%
                dplyr::select(-input_type, -unit)


info_tb_bind <- list()

# Loop through each year from 2016 to 2023
for (i in 1:8 ) {
  # Filter info_tb_102 for the current year
  info_tb_year <- info_tb_101[year == i+2015] 
  
  # Generate the file paths based on ffy for the current year
  all_ready_file_paths <- paste0(here("Data", "processed", "Analysis_ready"), "/", info_tb_year$ffy_id, "_merged_data.rds")
  
  # Read the RDS files for the current year
  comb_sf_list <- lapply(all_ready_file_paths, readRDS)

  mean_y_comb <- lapply(comb_sf_list, function(x) {
  mean(st_drop_geometry(x)$yield, na.rm = TRUE)
})
 
  info_tb_year$avg_yield <-  data.frame(mean_yield = unlist(mean_y_comb))

  info_tb_bind[[i]] <- info_tb_year

}

# data table that contains all the farm and experimental info by year 
info_tb_binded  <- bind_rows(info_tb_bind)

info_tb_binded <- info_tb_binded %>% filter(eosr_y >= 70)

dim(info_tb_binded)

# Step 1: Calculate the new different of seeding rate
# yield and profit by SQSR - EOSR 

info_tb_binded <- info_tb_binded[, `:=`(
  dif_s = sqsr - eosr,
  dif_y = sqsr_y - eosr_y,
  dif_p = sqsr_p - eosr_p)]


# Step 2: Calculate the quadratic fit and yield maximizing seeding rate
# for each field and Generate response type variable based on the conditions

info_tb_resp <- process_info_table(info_tb_binded)

saveRDS(info_tb_resp, here("Data", "processed", "Analysis_results", "info_tb_resp.rds"))


```

###### CONTENTS ######

### Tables ###

# table1  - Data summary by year (2016 to 2023) 
# [summary_flex_tab1] / [table1_dat_summary.png]

# table2  - Analysis results table By the response type and seeding status (SQSR-EOSR) 
# [results_flex_tab2] / [table2_results_summary.png]

# table3 - Discussion Table 
# [on-going]
# [What happens to the farmer's expected scenario in terms of profit, and how the problem going in the bad weather cases]

### Figures ###

# figure1  - Net revenue after seed cost figure (1996 to 2023)
# [rev_seed_comb] / [fig1_rev_after_seed.png]

# figure2 - trial_design sampple figure  
# [td_figure] / [fig2_trial_design_sample.png]

# figure3 - EOSR ~ Precipitation and GDD ( In-season & 30 year Avg)
# [eosr_sqsr_weather] / [fig3_eosr_sqsr_weather.png]

# figure4 - Difference in estimated Profit at a given climate condition by seeding status (SQSR-EOSR)
#  [dif_pro_seed_comb] / [fig5_dif_pro_seed_comb.png]

# figure5 - Estimated Profit and Yield response by type (A2,B2,C)
#  [type_all_fig] / [fig4_est_resp_type_all.png]



################################
######### Tables ###############
################################


## Table1 Data summary by year (2016 to 2023)
# [summary_tab1] / [table1_dat_summary.png]

```{r tables data table1, cache = T, results = "hide"}
info_tb_resp <-readRDS(here("Data", "processed", "Analysis_results", "info_tb_resp.rds"))


# Table 1 data summary table 


# Calculate mean, sd (with NA removal for usdsr), and field count by year
summary_tab1_convert <- info_tb_resp[, .(
  field_count = .N , # Count the number of rows for each year
  avg_yield = sprintf("%.1f\n(%.1f)", round(mean(avg_yield), 1), round(sd(avg_yield), 1)),
  sqsr = sprintf("%.1f\n(%.1f)", round(mean(sqsr), 1), round(sd(sqsr), 1)),
  usdsr = sprintf("%.1f\n(%.1f)", round(mean(usdsr, na.rm = TRUE), 1), round(sd(usdsr, na.rm = TRUE), 1)),
  prcp_t = sprintf("%.1f\n(%.1f)", round(mean(prcp_t), 1), round(sd(prcp_t), 1)),
  gdd_t = sprintf("%.1f\n(%.1f)", round(mean(gdd_t), 1), round(sd(gdd_t), 1)),
  prcp_30 = sprintf("%.1f\n(%.1f)", round(mean(prcp_30), 1), round(sd(prcp_30), 1)),
  gdd_30 = sprintf("%.1f\n(%.1f)", round(mean(gdd_30), 1), round(sd(gdd_30), 1))
), by = year]


summary_tab1 <- flextable(summary_tab1_convert) %>%
  set_header_labels(
    field_count = "Field\nCount",
    year = "Year", 
    avg_yield = "Mean\nYield\n(bu/ac)",
    sqsr = "SQSR",
    usdsr = "USDA\nSR", 
    prcp_t = "Precipitation\n(In-Season)",
    gdd_t = "GDD\n(In-Season)",
    prcp_30 = "Precipitation\n(30Year)",
    gdd_30 = "GDD\n(30Year)"
  ) %>%
  align(align = "center", part = "all") %>%  # Center-align headers and body content
  autofit() %>%
  set_table_properties(layout = "autofit")

# Print the flextable  (TABLE1)
summary_tab1

# Save the flextable as an HTML file
save_as_html(summary_tab1, path = here("Results","Tables","table1_dat_summary.html"))

# Convert the HTML to a PNG image
webshot(here("Results","Tables","table1_dat_summary.html"), 
        file = here("Results","Tables","table1_dat_summary.png"),
        zoom = 2)  # Adjust zoom if you need higher resoluti

```


## Table2 ( Results table By Response type and estimated profit ) 
# [results_flex_tab2] / [table2_results_summary.png]

```{r tables result table2 by response type a to d , cache = T, results = "hide"}


# Make results table by response type

result_table2 <- info_tb_resp[, .(
  field_count = .N,  # Count the number of rows for each resp_type
  dif_s = sprintf("%.1f\n(%.1f)", round(mean(dif_s), 1), round(sd(dif_s), 1)),
  dif_y = sprintf("%.1f\n(%.1f)", round(mean(dif_y), 1), round(sd(dif_y), 1)),
  dif_p = sprintf("%.1f\n(%.1f)", round(mean(dif_p), 1), round(sd(dif_p), 1)),
  prcp_t = sprintf("%.1f\n(%.1f)", round(mean(prcp_t), 1), round(sd(prcp_t), 1)),
  gdd_t = sprintf("%.1f\n(%.1f)", round(mean(gdd_t), 1), round(sd(gdd_t), 1)),
  prcp_30 = sprintf("%.1f\n(%.1f)", round(mean(prcp_30), 1), round(sd(prcp_30), 1)),
  gdd_30 = sprintf("%.1f\n(%.1f)", round(mean(gdd_30), 1), round(sd(gdd_30), 1))
), by = resp_type]

setorder(result_table2, resp_type)

# Convert to a flextable for a formatted summary
result_flex_tab2 <- flextable(result_table2) %>%
  set_header_labels(
    field_count = "Field\nCount",
    dif_s = "Differences in \n Seeding Rate\n(K/ac)",
    dif_y = "Differences in \n Estimated Yield \n (bu/ac)",
    dif_p = "Differences in \n Estimated Profit \n ($/ac)",
    prcp_t = "Precipitation\n(In-Season)",
    gdd_t = "GDD\n(In-Season)",
    prcp_30 = "Precipitation\n(30 Year)",
    gdd_30 = "GDD\n(30 Year)"
  ) %>%
  align(align = "center", part = "all") %>%  # Center-align headers and body content
  autofit() %>%
  set_table_properties(layout = "autofit")

# Print the flextable 
result_flex_tab2

# Save the flextable as an HTML file
save_as_html(result_flex_tab2, path = here("Results","Tables","table2_results_summary.html"))

# Convert the HTML to a PNG image
webshot(here("Results","Tables","table2_results_summary.html"), 
        file = here("Results","Tables","table2_results_summary.png"),
        zoom = 2)  # Adjust zoom if you need higher resoluti


```

##################################
############ Figures #############
##################################

## Figure 1  ( Net revenue after seed cost figure 1996 to 2023)
# [rev_seed_comb] / [fig1_rev_after_seed.png]

```{r figure1 net revenue after seed cost, cache = T, results = "hide"}

corn_ers <-read.csv(here("Data","Raw","corn_cost_return.csv"),header = T)
ppi_dat <-read.csv(here("Data","Raw","ppi_corn_annual.csv"),header = T)



# Prepare ppi data and calculate annual average
ppi_dat <- ppi_dat %>%
  mutate(DATE = as.Date(DATE), year = lubridate::year(DATE)) %>%
  group_by(year) %>%
  summarise(wpu_annual = mean(WPU01220205, na.rm = TRUE)) %>%
  filter(year >= 1996 & year <= 2023)

# Set 1996 PPI as baseline
wpu_1996 <- ppi_dat %>% filter(year == 1996) %>% pull(wpu_annual)

# Filter corn data
corn_sel <- corn_ers %>%
  filter(Item %in% c("Primary product, grain","Fertilizer  ", "Seed", "Yield"),
         Region == "U.S. total") %>%
  left_join(ppi_dat, by = c("Year" = "year")) %>%
  mutate(
    wpu_normalized = wpu_annual / wpu_1996,
    Value_adjusted = Value / wpu_normalized
  ) %>%
  drop_na()

saveRDS(corn_sel, here("Data", "Processed","usda_ers_corn_cost_return.rds"))


```


## Figure2 (Experimental design Sample)
# [td_figure] / [fig2_trial_design_sample.png]


```{r figure2 experimental design, cache = T, results = "hide"}

 td_sample <-st_read(here("Data","Raw","trial-design-sample.shp"))

# Mutating the trial_rate based on tgt_seed values
td_sample <- td_sample %>%
  mutate(trial_rate = case_when(
    tgt_seed == 130 ~ 32,
    tgt_seed == 145 ~ 30,
    tgt_seed == 110 ~ 28,
    tgt_seed == 90  ~ 34,
    tgt_seed == 160 ~ 36,
    TRUE ~ NA_real_  # If tgt_seed doesn't match any of the above values, assign NA
  )) %>% dplyr::select(trial_rate)

random_colors <- c("#E41A1C", "#377EB8", "#4DAF4A", "#984EA3", "#FF7F00")  # Example colors

td_figure  <- ggplot(td_sample) +
  geom_sf(aes(fill = factor(trial_rate))) +  # Convert trial_rate to a factor for discrete values
  scale_fill_manual(values = random_colors, name = "Trial Seed Rate (K/ac)") +  # Random color scale with custom legend title
  theme_minimal() +
  coord_sf(expand = FALSE) +  # Prevents automatic expansion of the plot
  theme(axis.title = element_blank(),   # Remove axis titles
        axis.text = element_blank(),    # Remove axis text
        axis.ticks = element_blank())   # Remove axis ticks

# Save the plot
ggsave(here("Results","Figures","fig2_trial_design_sample.png"), plot = td_figure, width = 6, height = 4)


```


# figure3 - EOSR ~ Precipitation ( In-season & 30 year Avg)
# [eosr_sqsr_weather] / [fig3_eosr_sqsr_weather.png]

```{r figure3 EOSR ~ Precipitation and GDD , cache = T, results = "hide"}

info_tb_resp <- readRDS(here("Data", "processed", "Analysis_results", "info_tb_resp.rds"))

names(info_tb_resp)

# Define min and max values for precipitation variables
x_range_prec <- range(info_tb_resp$prcp_t, na.rm = TRUE)
y_range <- range(c(info_tb_resp$eosr, info_tb_resp$sqsr))

# Create the first plot for in-season precipitation
eosr_sqsr_prec <- ggplot(info_tb_resp) +
  geom_point(aes(x = prcp_t, y = eosr, color = "EOSR")) +
  geom_point(aes(x = prcp_t, y = sqsr, color = "SQSR")) +
  scale_color_manual(name = "Type", values = c("EOSR" = "red", "SQSR" = "blue")) +
  xlim(200, 1000) +
  geom_abline(intercept = 0, slope = (diff(y_range) / diff(x_range_prec)), color = "black", size = 0.2) +
  labs(y = "Seeding Rate", x = "Precipitation \n In-Season (inch)") +
  theme_minimal(base_size = 14) +
  theme(legend.position = "none")

# Define min and max values for 30-year average precipitation
x_range_prcp30 <- range(info_tb_resp$prcp_30, na.rm = TRUE)

# Create the second plot for 30-year average precipitation
eosr_sqsr_prec30 <- ggplot(info_tb_resp) +
  geom_point(aes(x = prcp_30, y = eosr, color = "EOSR")) +
  geom_point(aes(x = prcp_30, y = sqsr, color = "SQSR")) +
  scale_color_manual(name = "Seeding Type", values = c("EOSR" = "red", "SQSR" = "blue")) +
  xlim(200, 1000) +
  geom_abline(intercept = 0, slope = (diff(y_range) / diff(x_range_prcp30)), color = "black", size = 0.2) +
  labs(x = "Precipitation \n (30-Year Average)", y = NULL) +
  theme_minimal(base_size = 14) +
  theme(
    legend.position = c(0.1, -0.3),
    legend.direction = "horizontal",
    legend.title = element_text(size = 14),
    legend.text = element_text(size = 14),
    plot.margin = margin(5, 5, 60, 5)
  )

# Combine the two plots
eosr_sqsr_weather <- (eosr_sqsr_prec | eosr_sqsr_prec30) +
  plot_annotation(
    theme = theme(plot.margin = margin(10, 10, 40, 10))
  )

# Save the combined plot
ggsave(here("Results", "Figures", "fig3_eosr_sqsr_weather.png"), plot = eosr_sqsr_weather, width = 16, height = 10, units = "in")

ggsave(here("Results","Figures","fig3_eosr_sqsr_weather.png"), plot =eosr_sqsr_weather , width = 16, height = 10, units = "in")

```


# figure4 - Difference in estimated Profit at a given climate condition by seeding status (SQSR-EOSR)

#  [dif_pro_seed_comb] / [fig4_dif_pro_seed_comb.png]

```{r figure4 - Difference in estimated Profit at a given climate condition by seeding status, cache = T, results = "hide"}

# Grouping by prcp_tot for dif_pro_seed
info_tb_resp$prcp_t_grp <- cut(
  info_tb_resp$prcp_t,
  breaks = quantile(info_tb_resp$prcp_t, probs = 0:10 / 10, na.rm = TRUE),
  include.lowest = TRUE
)

info_tb_resp$prcp_30_grp <- cut(
  info_tb_resp$prcp_30,
  breaks = quantile(info_tb_resp$prcp_30, probs = 0:10 / 10, na.rm = TRUE),
  include.lowest = TRUE
)

# Using colorRampPalette to create a  gradient
grp_cols_t<- colorRampPalette(c("#E5CCFF", "#A566FF", "#5C007F"))(10)  # Light to dark purple shades
grp_cols_30 <- colorRampPalette(c("#FFE5CC", "#FFCC66", "#FF6600"))(10)  # Light to dark orange shades

# Define labels for the legend based on numeric ranges
grp_t_labels <- levels(info_tb_resp$prcp_t_grp)
grp_30_labels <- levels(info_tb_resp$prcp_30_grp)

# Plot for prcp_t
dif_prof_seed_prec_t <- ggplot(info_tb_resp) +
  geom_point(aes(x = dif_s, y = dif_p, color = prcp_t_grp),size = 3) + scale_color_manual(
    name = "In-Season \n Precipitation (inch)",  # Legend title with two lines
    values = grp_cols_t,  # Colors for each prcp_group
    breaks = grp_t_labels,  # Display the actual ranges in the legend
    labels = grp_t_labels,  # Use numeric ranges as labels
     guide = guide_legend(ncol = 1)
  ) +
   geom_segment(aes(x = 0, y = -100, xend = 0, yend = 0), color = "black", size = 0.5,
                linetype='dashed') +  # Add vertical line
  scale_x_continuous(
    breaks = seq(floor(min(info_tb_resp$dif_s, na.rm = TRUE) / 5) * 5,
                 ceiling(max(info_tb_resp$dif_s, na.rm = TRUE) / 5) * 5,
                 by = 5)  # Set breaks every 5 units
  ) +
  ylim(-100, 0) +  # Set y-axis limits
  labs(
    x = NULL,
    y = "Estimated Profit Difference($/ac)"
  ) +
  theme(
    legend.position = "right", 
    legend.direction = 'vertical',# Place the legend at the bottom
      axis.text = element_text(size = 16),  # Increase axis text size
    axis.title = element_text(size = 16),  # Increase axis title size
    legend.title = element_text(size = 16),  # Increase legend title font size
    legend.text = element_text(size = 16)    # Increase legend text font size
  )


dif_prof_seed_prec_30 <- ggplot(info_tb_resp) +
  geom_point(aes(x = dif_s, y = dif_p, color = prcp_30_grp),size = 3) + scale_color_manual(
    name = "30 year average \n In-season \n Precipitation (inch)",  # Legend title with two lines
    values = grp_cols_30,  # Colors for each prcp_group
    breaks = grp_30_labels,  # Display the actual ranges in the legend
    labels = grp_30_labels,  # Use numeric ranges as labels
     guide = guide_legend(ncol = 1)
  ) +
   geom_segment(aes(x = 0, y = -100, xend = 0, yend = 0), color = "black", size = 0.5,
                linetype='dashed') +  # Add vertical line
  scale_x_continuous(
    breaks = seq(floor(min(info_tb_resp$dif_s, na.rm = TRUE) / 5) * 5,
                 ceiling(max(info_tb_resp$dif_s, na.rm = TRUE) / 5) * 5,
                 by = 5)  # Set breaks every 5 units
  ) +
  ylim(-100, 0) +  # Set y-axis limits
  labs(
    x = NULL,
    y = "Estimated Profit Difference($/ac)"
  ) +
  theme(
    legend.position = "right", 
    legend.direction = 'vertical',# Place the legend at the bottom
      axis.text = element_text(size = 16),  # Increase axis text size
    axis.title = element_text(size = 16),  # Increase axis title size
    legend.title = element_text(size = 16),  # Increase legend title font size
    legend.text = element_text(size = 16)    # Increase legend text font size
  )


# Combine the two plots side by side with patchwork
dif_prof_seed_prec_comb <- (dif_prof_seed_prec_t / dif_prof_seed_prec_30)   # Collect legends to appear only once


ggsave(here("Results","Figures","fig4_dif_prof_seed_prec_comb.png"), plot = dif_prof_seed_prec_comb , width = 14, height = 12, units = "in")


```

# figure5 - Estimated Profit and Yield response by type (A2,B2,C)
# [type_all_fig] / [fig5_resp_type_all.png]


```{r figure5 - Estimated Profit and Yield response by type  , cache = T, results = "hide"}

info_tb_resp <- readRDS(here("Data", "processed", "Analysis_results", "info_tb_resp.rds"))


info_tb_a2 <- info_tb_resp[resp_type == "A2"]
#ffy_id = "21_1_2023"

eval_type_a2 <-readRDS(here("Data", "processed", "Analysis_results","21_1_2023_eval_tb.rds"))

info_a2_chosen <-info_tb_a2[ffy_id == "21_1_2023",]

min_prof_lim_a2 = min(eval_type_a2$profit_hat_year) 
max_prof_lim_a2 = max(eval_type_a2$profit_hat_year)


  type_a2_p_resp <- ggplot(eval_type_a2, aes(x = s_rate, y = profit_hat_year)) +
  geom_smooth(method = "loess", color = 'yellow') +
  geom_point(data = info_a2_chosen, aes(x = sqsr, y = sqsr_p), color = "blue", size = 1.5) +
  geom_point(data = info_a2_chosen, aes(x = eosr, y = eosr_p), color = "red", size = 1.5) +
  geom_point(data = info_a2_chosen, aes(x = ymsr, y = ymsr_p), color = "black", size = 1.5)  +
 ylim(min_prof_lim_a2-0.05*(max_prof_lim_a2-min_prof_lim_a2),max_prof_lim_a2+ 0.05*(max_prof_lim_a2-min_prof_lim_a2)) +
 # Draw vertical lines with dashed linetype and specified y-axis limits
  geom_segment(data = info_a2_chosen, aes(x = sqsr, xend = sqsr, y = min(eval_type_a2$profit_hat_year), yend = sqsr_p), 
               linetype = "dashed", color = "blue") +
  geom_segment(data = info_a2_chosen, aes(x = eosr, xend = eosr, y = min(eval_type_a2$profit_hat_year), yend = eosr_p), 
               linetype = "dashed", color = "red") +
    geom_segment(data = info_a2_chosen, aes(x = ymsr, xend = ymsr, y = min(eval_type_a2$profit_hat_year), yend = ymsr_p), 
               linetype = "dashed", color = "black") +
               # Add text annotations for points
  geom_text(data = info_a2_chosen, aes(x = sqsr, y = min(eval_type_a2$profit_hat_year), 
                                           label = paste0("SQSR\n(", round(sqsr,1),"\nK/seed", ")")), color = 'blue',
            vjust = 2, hjust = 0.5, size = 4) +
    geom_text(data = info_a2_chosen, aes(x = eosr, y = min(eval_type_a2$profit_hat_year), 
                                           label = paste0("EOSR\n(", round(eosr,1),"\nK/seed", ")")), color = 'red',
            vjust = 2, hjust = 0.5, size = 4) +
     geom_text(data = info_a2_chosen, aes(x = ymsr, y = min(eval_type_a2$profit_hat_year), 
                                           label = paste0("YMSR\n(", round(ymsr,1),"\nK/seed", ")")), color = 'black',
            vjust = 2, hjust = 0.5, size = 4) +
             # Add value labels above points
  geom_text(data = info_a2_chosen, aes(x = sqsr, y = sqsr_p, label = paste0(round(sqsr_p - eosr_p  , 1),"\n($/ac)")), 
            color = 'blue', vjust = -1, hjust = 0.5, size = 4) +
  geom_text(data = info_a2_chosen, aes(x = ymsr, y = ymsr_p, label = paste0(round(ymsr_p - eosr_p , 1),"\n($/ac)")), 
            color = 'black', vjust = -1, hjust = 0.5, size = 4) +

  # Remove the legend
   theme(
    axis.title.x = element_text(size = 18),  # Increase font size for x-axis label
    axis.title.y = element_text(size = 18),  # Increase font size for y-axis label
    axis.text.x = element_text(size = 18),    # Increase font size for x-axis numeric values
    axis.text.y = element_text(size = 18) ,    # Increase font size for y-axis numeric values
     legend.position = "none") +
  
  # Customize labels
  labs( x = "Seeding Rate (K/ac)", y = "Estimated Revenue - Seed Cost ($/ac)")


## Add yield response plot side by side 
min_yield_lim_a2 = min(eval_type_a2$yield_hat) 
max_yield_lim_a2 = max(eval_type_a2$yield_hat)

 type_a2_y_resp <- ggplot(eval_type_a2, aes(x = s_rate, y = yield_hat)) +
  geom_smooth(method = "loess", color = 'green') +
  geom_point(data = info_a2_chosen, aes(x = sqsr, y = sqsr_y), color = "blue", size = 1.5) +
  geom_point(data = info_a2_chosen, aes(x = eosr, y = eosr_y), color = "red", size = 1.5) +
  geom_point(data = info_a2_chosen, aes(x = ymsr, y = ymsr_y), color = "black", size = 1.5)  +
  ylim(min_yield_lim_a2-0.05*(max_yield_lim_a2-min_yield_lim_a2),
  max_yield_lim_a2+0.05*(max_yield_lim_a2-min_yield_lim_a2)) +
 # Draw vertical lines with dashed linetype and specified y-axis limits
  geom_segment(data = info_a2_chosen, aes(x = sqsr, xend = sqsr, y = min(eval_type_a2$yield_hat), yend = sqsr_y), 
               linetype = "dashed", color = "blue") +
  geom_segment(data = info_a2_chosen, aes(x = eosr, xend = eosr, y = min(eval_type_a2$yield_hat), yend = eosr_y), 
               linetype = "dashed", color = "red") +
    geom_segment(data = info_a2_chosen, aes(x = ymsr, xend = ymsr, y = min(eval_type_a2$yield_hat), yend = ymsr_y), 
               linetype = "dashed", color = "black") +
  # Add text annotations for points
  geom_text(data = info_a2_chosen, aes(x = sqsr, y = min(eval_type_a2$yield_hat), 
                                       label = paste0("SQSR\n(", round(sqsr,1),"\nK/seed", ")")), color = 'blue',
            vjust = 2, hjust = 0.5, size = 4) +
  geom_text(data = info_a2_chosen, aes(x = eosr, y = min(eval_type_a2$yield_hat), 
                                       label = paste0("EOSR\n(", round(eosr,1),"\nK/seed", ")")), color = 'red',
            vjust = 2, hjust = 0.5, size = 4) +
  geom_text(data = info_a2_chosen, aes(x = ymsr, y = min(eval_type_a2$yield_hat), 
                                       label = paste0("YMSR\n(", round(ymsr,1),"\nK/seed", ")")), color = 'black',
            vjust = 2, hjust = 0.5, size = 4) +
  # Add value labels above points
  geom_text(data = info_a2_chosen, aes(x = sqsr, y = sqsr_y, label = paste0(round(eosr_y - sqsr_y, 1),"\n(bu/ac)")), 
            color = 'blue', vjust = -1, hjust = 0.5, size = 4) +
  geom_text(data = info_a2_chosen, aes(x = ymsr, y = ymsr_y, label = paste0(round(eosr_y - ymsr_y, 1),"\n(bu/ac)")), 
            color = 'black', vjust = -1, hjust = 0.5, size = 4) +
   theme(
    axis.title.x = element_text(size = 18),  # Increase font size for x-axis label
    axis.title.y = element_text(size = 18),  # Increase font size for y-axis label
    axis.text.x = element_text(size = 18),    # Increase font size for x-axis numeric values
    axis.text.y = element_text(size = 18) ,    # Increase font size for y-axis numeric values
     legend.position = "none") +
  
  # Customize labels
  labs( x = "Seeding Rate (K/ac)", y = "Yield (bu/ac)")

### Make the plot side by side

type_a2_fig <- type_a2_p_resp | type_a2_y_resp



####################################################################
###################Type B2    ######################################
####################################################################

info_tb_b2 <- info_tb_resp[resp_type == "B2"]

info_tb_b2$dif_y
info_tb_b2$ffy_id[39]
#ffy_id = "5_1_2023"

eval_type_b2 <-readRDS(here("Data", "processed", "Analysis_results","5_1_2023_eval_tb.rds"))

info_b2_chosen <-info_tb_b2[ffy_id == "5_1_2023",]

min_prof_lim_b2 = min(eval_type_b2$profit_hat_year) 
max_prof_lim_b2 = max(eval_type_b2$profit_hat_year)

  type_b2_p_resp <- ggplot(eval_type_b2, aes(x = s_rate, y = profit_hat_year)) +
  geom_smooth(method = "loess", color = 'yellow') +
  geom_point(data = info_b2_chosen, aes(x = sqsr, y = sqsr_p), color = "blue", size = 1.5) +
  geom_point(data = info_b2_chosen, aes(x = eosr, y = eosr_p), color = "red", size = 1.5) +
  geom_point(data = info_b2_chosen, aes(x = ymsr, y = ymsr_p), color = "black", size = 1.5)  +
 ylim(min_prof_lim_b2-0.05*(max_prof_lim_b2-min_prof_lim_b2),max_prof_lim_b2+ 0.05*(max_prof_lim_b2-min_prof_lim_b2)) +
 # Draw vertical lines with dashed linetype and specified y-axis limits
  geom_segment(data = info_b2_chosen, aes(x = sqsr, xend = sqsr, y = min(eval_type_b2$profit_hat_year), yend = sqsr_p), 
               linetype = "dashed", color = "blue") +
  geom_segment(data = info_b2_chosen, aes(x = eosr, xend = eosr, y = min(eval_type_b2$profit_hat_year), yend = eosr_p), 
               linetype = "dashed", color = "red") +
    geom_segment(data = info_b2_chosen, aes(x = ymsr, xend = ymsr, y = min(eval_type_b2$profit_hat_year), yend = ymsr_p), 
               linetype = "dashed", color = "black") +
               # Add text annotations for points
  geom_text(data = info_b2_chosen, aes(x = sqsr, y = min(eval_type_b2$profit_hat_year), 
                                           label = paste0("SQSR\n(", round(sqsr,1),"\nK/seed", ")")), color = 'blue',
            vjust = 2, hjust = 0.5, size = 4) +
    geom_text(data = info_b2_chosen, aes(x = eosr, y = min(eval_type_b2$profit_hat_year), 
                                           label = paste0("EOSR\n(", round(eosr,1),"\nK/seed", ")")), color = 'red',
            vjust = 2, hjust = 0.5, size = 4) +
     geom_text(data = info_b2_chosen, aes(x = ymsr, y = min(eval_type_b2$profit_hat_year), 
                                           label = paste0("YMSR\n(", round(ymsr,1),"\nK/seed", ")")), color = 'black',
            vjust = 2, hjust = 0.5, size = 4) +
             # Add value labels above points
  geom_text(data = info_b2_chosen, aes(x = sqsr, y = sqsr_p, label = paste0(round(sqsr_p - eosr_p  , 1),"\n($/ac)")), 
            color = 'blue', vjust = -1, hjust = 0.5, size = 4) +
  geom_text(data = info_b2_chosen, aes(x = ymsr, y = ymsr_p, label = paste0(round(ymsr_p - eosr_p , 1),"\n($/ac)")), 
            color = 'black', vjust = -1, hjust = 0.5, size = 4) +
  
  
  # Remove the legend
   theme(
    axis.title.x = element_text(size = 18),  # Increase font size for x-axis label
    axis.title.y = element_text(size = 18),  # Increase font size for y-axis label
    axis.text.x = element_text(size = 18),    # Increase font size for x-axis numeric values
    axis.text.y = element_text(size = 18) ,    # Increase font size for y-axis numeric values
     legend.position = "none") +
  
  # Customize labels
  labs( x = "Seeding Rate (K/ac)", y = "Estimated Revenue - Seed Cost ($/ac)")

## Add yield response plot side by side 
min_yield_lim_b2 = min(eval_type_b2$yield_hat) 
max_yield_lim_b2 = max(eval_type_b2$yield_hat)

 type_b2_y_resp <- ggplot(eval_type_b2, aes(x = s_rate, y = yield_hat)) +
  geom_smooth(method = "loess", color = 'green') +
  geom_point(data = info_b2_chosen, aes(x = sqsr, y = sqsr_y), color = "blue", size = 1.5) +
  geom_point(data = info_b2_chosen, aes(x = eosr, y = eosr_y), color = "red", size = 1.5) +
  geom_point(data = info_b2_chosen, aes(x = ymsr, y = ymsr_y), color = "black", size = 1.5)  +
  ylim(min_yield_lim_b2-0.05*(max_yield_lim_b2-min_yield_lim_b2),max_yield_lim_b2+0.05*(max_yield_lim_b2-min_yield_lim_b2)) +
 # Draw vertical lines with dashed linetype and specified y-axis limits
  geom_segment(data = info_b2_chosen, aes(x = sqsr, xend = sqsr, y = min(eval_type_b2$yield_hat), yend = sqsr_y), 
               linetype = "dashed", color = "blue") +
  geom_segment(data = info_b2_chosen, aes(x = eosr, xend = eosr, y = min(eval_type_b2$yield_hat), yend = eosr_y), 
               linetype = "dashed", color = "red") +
    geom_segment(data = info_b2_chosen, aes(x = ymsr, xend = ymsr, y = min(eval_type_b2$yield_hat), yend = ymsr_y), 
               linetype = "dashed", color = "black") +
  # Add text annotations for points
  geom_text(data = info_b2_chosen, aes(x = sqsr, y = min(eval_type_b2$yield_hat), 
                                       label = paste0("SQSR\n(", round(sqsr,1),"\nK/seed", ")")), color = 'blue',
            vjust = 2, hjust = 0.5, size = 4) +
  geom_text(data = info_b2_chosen, aes(x = eosr, y = min(eval_type_b2$yield_hat), 
                                       label = paste0("EOSR\n(", round(eosr,1),"\nK/seed", ")")), color = 'red',
            vjust = 2, hjust = 0.5, size = 4) +
  geom_text(data = info_b2_chosen, aes(x = ymsr, y = min(eval_type_b2$yield_hat), 
                                       label = paste0("YMSR\n(", round(ymsr,1),"\nK/seed", ")")), color = 'black',
            vjust = 2, hjust = 0.5, size = 4) +
  # Add value labels above points
  geom_text(data = info_b2_chosen, aes(x = sqsr, y = sqsr_y, label = paste0(round(eosr_y - sqsr_y, 1),"\n(bu/ac)")), 
            color = 'blue', vjust = -1, hjust = 0.5, size = 4) +
  geom_text(data = info_b2_chosen, aes(x = ymsr, y = ymsr_y, label = paste0(round(eosr_y - ymsr_y, 1),"\n(bu/ac)")), 
            color = 'black', vjust = -1, hjust = 0.5, size = 4) +
   theme(
    axis.title.x = element_text(size = 18),  # Increase font size for x-axis label
    axis.title.y = element_text(size = 18),  # Increase font size for y-axis label
    axis.text.x = element_text(size = 18),    # Increase font size for x-axis numeric values
    axis.text.y = element_text(size = 18) ,    # Increase font size for y-axis numeric values
     legend.position = "none") +
  
  # Customize labels
  labs( x = "Seeding Rate (K/ac)", y = "Yield (bu/ac)")

### Make the plot side by side

type_b2_fig <- type_b2_p_resp | type_b2_y_resp

####################################################################
# Place type_a2_fig and type_b2_fig up and down

# Add captions with customized theme settings
type_a2_fig <- type_a2_fig +
  labs(caption = "Caption for type_a2_fig") +
  theme(
    plot.caption = element_text(size = 14, hjust = 0.5, margin = margin(t = 10, b = 10, l = 20, r = 20))
  )

type_b2_fig <- type_b2_fig +
  labs(caption = "Caption for type_b2_fig") +
  theme(
    plot.caption = element_text(size = 14, hjust = 0.5, margin = margin(t = 10, b = 10, l = 20, r = 20))
  )


# Combine the plots
type_all_fig <- type_a2_fig / type_b2_fig  
     
ggsave(here("Results","Figures","fig5_est_resp_type_all.png"),plot=type_all_fig, width = 16, height = 16, units = "in")

```
